add lonestar run scripts for is2, lj

+++ prog/hist2.h
@@ -358,7 +358,7 @@ __inline static int hist2_load(hist2_t *hs, const char *fn, unsigned flags)
       }
       if ( !hashist ) g = g2 * fac;
       if ( flags & HIST2_INT ) {
-        g = (long) (g + 0.5);
+        g = (double) (long) (g + 0.5);
       }
       if ( add ) g += hs->arr[r*nm + i*m + j];
       hs->arr[r*nm + i*m + j] = g;
